Use function local imports to eliminate top-level dependencies in Runtime
